refactor: extract own context menu module
from UnicodeFontHandler, which now solely handles the font/transformation stuff
