+++ src/background/background.js
@@ -1,6 +1,6 @@
 import * as AutocorrectHandler from "/common/modules/AutocorrectHandler.js";
-import * as UnicodeFontHandler from "/common/modules/UnicodeFontHandler.js";
+import * as ContextMenu from "./modules/ContextMenu.js";
 
 // init modules
 AutocorrectHandler.init();
-UnicodeFontHandler.init();
+ContextMenu.init();
